vat tests

+++ BusinessSimulation.Model/IVat.cs
@@ -6,8 +6,8 @@ namespace BusinessSimulation.Impl
 {
     public interface IVat
     {
-        float percent { get; set; }
+        double percent { get; set; }
 
-        float calculateTTC(float amount);
+        double calculateTTC(double amount);
     }
 }
